requirements.txt
django
scikit-learn
numpy
requests
gunicorn
whitenoise
python-dotenv
openai
sentence-transformers